backend/requirements.txt
asgiref==3.11.1
bcrypt==4.1.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.11
pycparser==3.0
python-dotenv==1.0.0
PyJWT==2.12.1
pytz==2026.1.post1
requests==2.31.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3